feat: add gRPC status code formatting utility for improved logging
This commit introduces a new header file, `grpcstatuscodefmt.h`, which provides a utility function to convert gRPC status codes to string representations. Additionally, it updates the logging in `remoteinputlistener.cpp` to utilize the new status code formatting, enhancing the clarity of error messages in the logging system. These changes improve the maintainability and readability of gRPC-related logs.

+++ evrp/device/client/remoteinputlistener.cpp
@@ -3,6 +3,7 @@
 #include <google/protobuf/empty.pb.h>
 
 #include "evrp/device/common/devicesessionmetadata.h"
+#include "evrp/device/internal/grpcstatuscodefmt.h"
 #include "evrp/device/internal/tofromproto.h"
 #include "logger.h"
 
@@ -34,7 +35,7 @@ bool RemoteInputListener::startListening(
   if (!st.ok()) {
     logError(
         "RemoteInputListener: StartRecording gRPC code={} {}",
-        static_cast<int>(st.error_code()),
+        st.error_code(),
         st.error_message());
     return false;
   }
